Baha c

See merge request chayma.saidi/bcm_back!31

+++ src/main/java/com/Bcm/Service/Impl/ProductOfferingServiceImpl/SubClassesServiceImpl/CategoryServiceImpl.java
@@ -4,6 +4,7 @@ import com.Bcm.Model.ProductOfferingABE.SubClasses.Category;
 import com.Bcm.Repository.ProductOfferingRepo.SubClassesRepo.CategoryRepository;
 import com.Bcm.Service.Srvc.ProductOfferingSrvc.SubClassesSrvc.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,7 +18,16 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category create(Category category) {
-        return categoryRepository.save(category);
+        try {
+            if (category == null) {
+                throw new IllegalArgumentException("Category cannot be null");
+            }
+            return categoryRepository.save(category);
+        } catch (DataIntegrityViolationException e) {
+            throw new IllegalArgumentException("There was a Data integrity violation occurred: " + e.getMessage());
+        } catch (Exception e) {
+            throw new RuntimeException("Error occurred while creating category: " + e.getMessage());
+        }
     }
 
     @Override
@@ -27,21 +37,32 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category update(int po_CategoryCode, Category updatedCategory) {
-        Optional<Category> existingCategoryOptional = categoryRepository.findById(po_CategoryCode);
+        try {
+            Optional<Category> existingCategoryOptional = categoryRepository.findById(po_CategoryCode);
 
-        if (existingCategoryOptional.isPresent()) {
-            Category existingCategory = existingCategoryOptional.get();
-            existingCategory.setName(updatedCategory.getName());
-            return categoryRepository.save(existingCategory);
-        } else {
-            throw new RuntimeException("Could not find Category with ID: " + po_CategoryCode);
+            if (existingCategoryOptional.isPresent()) {
+                Category existingCategory = existingCategoryOptional.get();
+                existingCategory.setName(updatedCategory.getName());
+                return categoryRepository.save(existingCategory);
+            } else {
+                throw new RuntimeException("Could not find Category with ID: " + po_CategoryCode);
+            }
+        } catch (IllegalArgumentException ex) {
+            throw new IllegalArgumentException("Invalid Category Code: " + po_CategoryCode);
+        } catch (Exception ex) {
+            throw new RuntimeException("An error occurred while updating the Category: " + ex.getMessage());
         }
     }
 
     @Override
     public String delete(int po_CategoryCode) {
-        categoryRepository.deleteById(po_CategoryCode);
-        return ("Category was successfully deleted");
+        Optional<Category> optionalCategory = categoryRepository.findById(po_CategoryCode);
+        if (optionalCategory.isPresent()) {
+            categoryRepository.deleteById(po_CategoryCode);
+            return ("Category was successfully deleted");
+        } else {
+            throw new RuntimeException("Could not find Category with ID: " + po_CategoryCode);
+        }
     }
 
     @Override
